working background session / manager
this is a bit overly optimistic...it assumes that once the upload task has been created
and the temporary file has been written that it will *eventually* get to the server

the in-memory store is then purged of the cached data.

this is not very fault-tolerant...but want to try something.