presenting an alert if username contains e-mail address. bumping version to 1.8.2
the username used by the fluxtream  API seems to be separate from the e-mail address
used by the fluxtream.com login. this causes endless confusion as there is no
immediate username verification when it is changed and failed authorization
appears on the main heart rate capture tab.

this change hopes to alleviate the confusion by warning users
that an e-mail address is likely incorrect, though it will save the
entered username value.